+++ R/google_distance.R
@@ -4,10 +4,12 @@
 #' and time for a matrix of origins and destinations, based on the recommended
 #' route between start and end points.
 #'
-#' @param origins list of unnamed elements, each element is either a numeric
-#' vector of lat/lon coordinates, or an address string
-#' @param destinations list of unnamed elements, each element is either a vector
-#' of lat/lon coordinates, or an address string
+#' @param origins Origin locations as either a one or two column data.frame, a
+#' list of unnamed elements, each element is either a numeric vector of lat/lon
+#' coordinates, or an address string, or a vector of a pair of lat / lon coordinates
+#' @param destinations destination locations as either a one or two column data.frame, a
+#' list of unnamed elements, each element is either a numeric vector of lat/lon
+#' coordinates, or an address string, or a vector of a pair of lat / lon coordinates
 #' @param mode \code{string} One of 'driving', 'walking', 'bicycling' or 'transit'.
 #' @param departure_time \code{POSIXct}. Specifies the desired time of departure. Must
 #' be in the future (i.e. greater than \code{sys.time()}). If no value is specified
@@ -37,14 +39,22 @@
 #' @return Either list or JSON string of the distance between origins and destinations
 #' @examples
 #' \dontrun{
+#'
+#' api_key <- 'your_api_key'
 #' google_distance(origins = list(c("Melbourne Airport, Australia"),
 #'                              c("MCG, Melbourne, Australia"),
 #'                              c(-37.81659, 144.9841)),
 #'                              destinations = c("Portsea, Melbourne, Australia"),
-#'                              key = "<your valid api key>",
+#'                              key = api_key,
 #'                              simplify = FALSE)
 #'
+#' google_distance(origins = c(-37.816, 144.9841),
+#'     destinations = c("Melbourne Airport, Australia", "Flinders Street Station, Melbourne"),
+#'     key = api_key)
 #'
+#' google_distance(origins = tram_stops[1:5, c("stop_lat", "stop_lon")],
+#'      destinations = tram_stops[10:12, c("stop_lat", "stop_lon")],
+#'      key = api_key)
 #'
 #' }
 #' @export
@@ -62,6 +72,9 @@ google_distance <- function(origins,
                             key,
                             simplify = TRUE,
                             curl_proxy = NULL){
+
+  origins <- validateLocations(origins)
+  destinations <- validateLocations(destinations)
 
   directions_data(base_url = "https://maps.googleapis.com/maps/api/distancematrix/json?",
                 information_type = "distance",
@@ -85,3 +98,37 @@ google_distance <- function(origins,
                 curl_proxy)
 
 }
+
+validateLocations <- function(locations) UseMethod("validateLocations")
+
+# validateLocations.list <- function(locations) locations
+#
+#' @export
+validateLocations.character <- function(locations) {
+
+  locations
+
+}
+
+#' @export
+validateLocations.numeric <- function(locations) {
+  ## a vector has to be put into a list
+  if(length(locations) > 2) stop("A vector can have a maximum of two elements")
+  list(locations)
+}
+
+#' @export
+validateLocations.data.frame <- function(locations){
+
+  ## A dataframe can be used, and can be one column or two
+  ##
+  if(ncol(locations) > 2) stop("A data.frame can have a maximum of two columns")
+  locations <- lapply(1:nrow(locations), function(x) as.numeric(locations[x, ]))
+  return(locations)
+}
+
+#' @export
+validateLocations.default <- function(locations) locations
+
+
+