+++ NEWS.md
@@ -1,5 +1,6 @@
 ## Version 2.3.0 (development)
 
+* `google_distance()` now accepts a data.frame of locations
 * `drag_drop_geojson` drag & drop geojson onto a map
 * `decode_pl` updated to now handle `jqr`-parsed character vectors
 * `colours` - You can now map variables passed into the map layers to colours